requirements.txt
# OAuth and HTTP Server
authlib==1.3.0
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# MCP SDK
mcp>=0.9.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
cryptography==41.0.7
pydantic==2.5.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

